Fix(Hypernative): refactoring for No Balance Check
