+++ apps/web/src/features/hypernative/hooks/index.ts
@@ -2,6 +2,6 @@ export { useIsHypernativeGuard } from './useIsHypernativeGuard'
 export type { HypernativeGuardCheckResult } from './useIsHypernativeGuard'
 export { useIsHypernativeFeature as useIsHypernativeFeatureEnabled } from './useIsHypernativeFeature'
 export { useBannerStorage } from './useBannerStorage'
-export type { BannerType } from './useBannerStorage'
+export { BannerType } from './useBannerStorage'
 export { useBannerVisibility, MIN_BALANCE_USD } from './useBannerVisibility'
 export type { BannerVisibilityResult } from './useBannerVisibility'
